Project/Sentimental_Analysis_with_Twitter_Data/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.1.1
future==0.18.2
idna==2.10
joblib==0.16.0
kafka-python==2.0.1
nltk==3.5
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.1
PySocks==1.7.1
python-dateutil==2.8.1
python-twitter==3.5
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
textblob==0.15.3
tqdm==4.48.2
tweepy==3.9.0
urllib3==1.25.10
